requirements.txt
acres==0.2.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
etelemetry==0.3.1
executing==2.2.0
filelock==3.17.0
fitz==0.0.1.dev2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.32.0
isodate==0.6.1
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
looseversion==1.3.0
lxml==5.3.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.3
openai==1.63.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
prov==2.0.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.1
pyxnat==1.6.3
pyzmq==26.2.1
rdflib==6.3.2
requests==2.32.3
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
